break up wallet source into multiple files, to make it easier to find specific code
